Rearrange and new files
I have created new header files so we do not have everything in the main.cpp

I have implemented the scheme in the case that the parameter satirfy the inequality of proposition 2.1. of the article

This is done in CIR2.h, as you will see the CIR2 function can be completes in the case that that inequality is not satisfy, then we need to define the threshold and the different schemes depending in the initial value x prior sampling.

In Distributions.h we can define the necesary functions to do the samplings.

Plots.h is for plots lol

You will also find that I have started a Python notebook to plot some samples from the second order scheme, you only need to change the path to the one in which the csv file is going to be stored

Feel free to rearrenge, the C++ files also have comments