Changes in functions and new modules
Notice that the new functions cir2_heston at cir2.cpp and exact_heston at schemes.cpp are the functions that given a vector of samples from a normal r.v. will return a path of volatilities. Those functions are the ones we want to use in the heston model :)

The rest of new modules and new functions are not well commented but I mostly need them to reproduce the convergence plots, I would comment them once I have some spare time